[ASSETS_LUG_V2] target platform assets LUT is copied to Resources only if it exists.
Former-commit-id: c94f8b721d17109aac787057a1cd2bb2d371db8f

+++ Assets/UbiPackages/Addressables/Editor/EditorAddressablesMenu.cs
@@ -248,9 +248,16 @@ public class EditorAddressablesMenu : MonoBehaviour
         }
 
         assetsLUTSource += ".json";
-        
-        File.Copy(assetsLUTSource, assetsLUTInResources);
+
+        if (File.Exists(assetsLUTSource))
+        {
+            File.Copy(assetsLUTSource, assetsLUTInResources);
         }
+        else
+        {
+            Debug.LogWarning("No assetsLUT found for platform " + target.ToString() + ": " + assetsLUTSource);
+        }
+
         AssetDatabase.Refresh();
     }
